blog/requirements.txt
amqp==2.2.2
anyjson==0.3.3
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
Django==2.0.1
django-appconf==1.0.2
django-celery==3.2.2
django-celery-email==2.0.0
django-celery-ses==0.9.2
django-crispy-forms==1.7.2
django-markdown-deux==1.0.5
django-newsletter-subscription==0.4.0
django-pagedown==1.0.4
django-widget-tweaks==1.4.1
idna==2.6
kombu==4.1.0
Markdown==2.6.9
markdown2==2.3.5
oauthlib==2.0.6
Pillow==5.0.0
PyJWT==1.5.3
python-decouple==3.1
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
urllib3==1.22
vine==1.1.4
